Fix macOS build: resolve ambiguous operator and type errors on ARM
On ARM macOS, size_t is unsigned long while uint64_t is unsigned long long
(both 64-bit but distinct types), causing several build failures:

- Ambiguous operator< between static_val<uint64_t> and size_type: add
  explicit static_cast<uint64_t>() on .size() calls in for-loop bounds
- std::clamp/std::max type deduction failures: use static_cast<uint64_t>()
  or explicit uint64_t type instead of UL suffix literals
- ConfigParameter<size_t> not in variant type list: change to uint64_t
- val<size_t&> not convertible to val<uint64_t>: change member type to
  uint64_t and read via getMemberWithOffset<uint64_t>
- DOMAIN macro conflict with macOS math.h: rename to SOCKET_DOMAIN

+++ nes-nautilus/src/Nautilus/Interface/BufferRef/ColumnTupleBufferRef.cpp
@@ -58,7 +58,7 @@ Record ColumnTupleBufferRef::readRecord(
 {
     Record record;
     const auto bufferAddress = recordBuffer.getMemArea();
-    for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
     {
         const auto& [name, type, dataTypeSize, columnOffset] = fields.at(i);
         if (not includesField(projections, name))
@@ -84,7 +84,7 @@ TupleBufferRef::WriteRecordResult ColumnTupleBufferRef::writeRecord(
     if (recordIndex < capacity)
     {
         const auto bufferAddress = recordBuffer.getMemArea();
-        for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+        for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
         {
             const auto& [name, type, dataTypeSize, columnOffset] = fields.at(i);
             if (not rec.hasField(name))
